requirements.txt
Django
djangorestframework
psycopg2
pre-commit
drf-spectacular
uuid
djangorestframework-simplejwt
